Move kafka-config map watcher to common pkg
